add new core option : "Target FPS for stutter reduction"
Inspired from this commit made by StupidHoroscope :
StupidHoroscope/libretro-scummvm-miyoo-backend@6c3954f

It allows to run some heavy games with less audio stuttering on the Miyoo Mini.

Default value is 55fps and correspond to an audio buffer of about 800 set it to 20 or 30 fps to obtain a buffer of 2000 which helps with heavy games in low power devices like Grim Fandango or Borken Sword 2.5.